Add files via upload
This script was written for exercise number 02 on Pentester Lab's Essential Badge.

Had to kinda bruteforce a password match w/ regex in a URL.

It's simple, but I take pride in it, as it took me a long time to get it to work. 

Uploading as a reminder for myself.